Added methods createStepExpected & createStepActual
-updates to writeReportToHtml to create report when method throws
exception
-added new methods to create step details for report.
-changed intLoop which has not been implemented to hold a integer
milliseconds to wait after an element is set in elementSet to handle
situations with Postbacks and the need to wait on some steps.  Default
value is 0 add integer in milliseconds to json parameter to sleep step
after set is performed.   This will most likely be permanent and require
a change or addition in the name from intLoop to
intMillisecondsWaitAfterSet.